Add files via upload
In the dynamic landscape of cybersecurity, organizations face a multitude of threats, and among them, Business Email Compromise (BEC) stands out as a pervasive and highly sophisticated attack vector. In response to the escalating threat posed by BEC incidents, I have developed a comprehensive BEC Playbook attached to this post. This playbook serves as a strategic guide to equip different organizations around the world with the knowledge and tools needed to detect, respond to, and mitigate the risks associated with BEC attacks.
This document is designed and developed by drawing insights from various regulatory standards, including but not limited to the ISO 27000 series, NIST standards, PCI DSS, and several others.


Should you have any inquiries, questions or concerns, please do not hesitate to contact me.

Your feedback is greatly valued and appreciated!!

#cyberdefense#cyberattack#cyberawareness#cybereducation#cyberdefenders#incidentresponse#incidentmanagement#incidentinvestigation#bec#businessemailcompromise#riskmanagement#riskanalysis#vulnerabilities#vulnerabilitymanagement#organizationaldevelopment#cybersecurityawareness#isaca#isc2#isc2congress#cybersecuritytraining#cybereducation#cybersecurity#publication#informationsecuritycommunity#informationsecurity#informationtechnology#informationsecurityawareness#informationsecuritymanagement#privacyprotection